figured out how to rotate a tool!
have to decouple the invert transform for PlayerBot though, so when the sprite flips the whole object does and the child VacStick object points the wrong way. Also added a little sprite!
